+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Now when an user is authenticated all the URLs created are linked to the user.
 - New route to get all the URLs created by the user.
+- New route to delete a URL created by the user.
+- New route to update a URL created by the user.
 
 ### Changed